Add test for hpack_huffman_encode

+++ src/hpack_huffman.c
@@ -857,9 +857,9 @@ const huffman_tree_t huffman_tree = {
  * - *huffman_tree: Pointer to the huffman tree storing the encoded symbols
  * - *result: Pointer to struct to save the result of the encoding
  * - sym: Symbol to encode
- * Output: 0 if it can encode the given symbol, 1 otherwise
+ * Output: 0 if it can encode the given symbol, -1 otherwise
  */
-int8_t hpack_huffman_encode(const huffman_tree_t *huffman_tree, huffman_encoded_word_t *result, uint8_t sym)
+int8_t hpack_huffman_encode(const huffman_tree_t *huffman_tree, huffman_encoded_word_t *result, uint16_t sym)
 {
     result->code = huffman_tree->codes[sym];
 #ifdef INCLUDE_HUFFMAN_LENGTH_TABLE
