+++ src/hpack_huffman.h
@@ -5,7 +5,7 @@
 #ifndef HPACK_HUFFMAN_H
 #define HPACK_HUFFMAN_H
 
-#include <stdio.h>
+#include <stdint.h>
 
 #define HUFFMAN_TABLE_SIZE 257
 #define NUMBER_OF_CODE_LENGTHS 21
@@ -26,7 +26,7 @@ typedef struct {
     const uint8_t code_length[NUMBER_OF_CODE_LENGTHS];
 } huffman_tree_t;
 
-int8_t hpack_huffman_encode(const huffman_tree_t *huffman_tree, huffman_encoded_word_t *result, uint8_t sym);
+int8_t hpack_huffman_encode(const huffman_tree_t *huffman_tree, huffman_encoded_word_t *result, uint16_t sym);
 
 /*uint8_t huffman_decode(const huffman_tree_t *huffman_tree, huffman_encoded_word_t *encoded)*/
